+++ lib/protocol/multipart/headers.rb
@@ -5,17 +5,24 @@
 
 require "protocol/http/headers"
 
+require_relative "header/content_disposition"
+require_relative "header/content_type"
+
 module Protocol
 	module Multipart
+		# @namespace
+		module Header
+		end
+		
 		# The header fields associated with a multipart body or part.
 		class Headers < Protocol::HTTP::Headers
 			POLICY = {
 				"content-description" => false,
-				"content-disposition" => false,
+				"content-disposition" => Header::ContentDisposition,
 				"content-id" => false,
 				"content-length" => false,
 				"content-transfer-encoding" => false,
-				"content-type" => false,
+				"content-type" => Header::ContentType,
 				"mime-version" => false,
 			}.tap do |policy|
 				policy.default = Protocol::HTTP::Header::Multiple